#Breaking Change
I would add the minimum HA-Version of 2021.9.0 so you are not on a path that still has a critical vurnability and also are on the same minimum version as HACS.
Also remove the Country stuff(Why was it in there?
And remove render_readme since we have a new info.md.